hms/requirements.txt
Django==4.2.28
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
django-redis==5.4.0
gunicorn==23.0.0
python-dotenv==1.0.1
Pillow==11.1.0
